+++ main.go
@@ -1,11 +1,9 @@
 package main
 
-import "os"
-
 const (
 	VERSION = "1.1.0"
 )
 
 func main() {
-	RunInterpreter(os.Args)
+	RunInterpreter()
 }